requirements.txt
web3
tweepy
asyncpg
nanoid
python-dotenv